Revert "add lido gateway / fetch l1gateway if not included (#31)"
This reverts commit b55729bd8744769c5e120701ee1f36867027ccc2.
